# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/alias": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alias"
                ],
                "summary": "Create alias",
                "parameters": [
                    {
                        "description": "Alias request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AliasReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRes"
                        }
                    }
                }
            }
        },
        "/alias/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get alias by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alias"
                ],
                "summary": "Get alias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alias ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Alias"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRes"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alias"
                ],
                "summary": "Update alias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alias ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Alias request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AliasReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRes"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alias"
                ],
                "summary": "Delete alias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alias ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRes"
                        }
                    }
                }
            }
        },
        "/aliases": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all aliases",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alias"
                ],
                "summary": "Get aliases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alias"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AliasReq": {
            "type": "object",
            "required": [
                "domain",
                "recipients"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "format": {
                    "type": "string"
                },
                "from_name": {
                    "type": "string"
                },
                "recipients": {
                    "type": "string"
                }
            }
        },
...
        "api.ErrorRes": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessRes": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Alias": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "from_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recipients": {
                    "type": "string"
                },
                "stats": {
                    "$ref": "#/definitions/model.AliasStats"
                }
            }
        },
        "model.AliasStats": {
            "type": "object",
            "properties": {
                "bandwidth": {
                    "type": "integer"
                },
                "blocks": {
                    "type": "integer"
                },
                "forwards": {
                    "type": "integer"
                },
                "replies": {
                    "type": "integer"
                },
                "sends": {
                    "type": "integer"
                }
            }
        },
...
    }
}